# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Obiara API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Authoritative REST contract for Obiara member clients. Durable API changes
    are additive within v1. Retriable writes require an Idempotency-Key.
...
servers:
  - url: /
    description: API root
...
paths:
...
  /v1/escrows:
    get:
      operationId: listMemberEscrows
      tags: [Commerce]
      summary: List the authenticated member's funded escrows
      security:
        - MemberBearer: []
      responses:
        "200":
          description: Member-owned escrow projections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowListEnvelope"
        "401":
          description: A valid member session is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Escrow service unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/escrows/{id}:
    get:
      operationId: getMemberEscrow
      tags: [Commerce]
      summary: Get one member-owned escrow
      security:
        - MemberBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Escrow projection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowEnvelope"
        "401":
          description: A valid member session is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Escrow not found or not owned by this member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/escrows/{id}/milestones/{milestoneId}/acceptance:
    post:
      operationId: acceptEscrowMilestone
      tags: [Commerce]
      summary: Record member acceptance evidence
      description: The authenticated member can record acceptance only. Matchmaker delivery and finance settlement are separate authorities.
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: milestoneId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "200":
          description: Acceptance evidence retained.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          description: A valid member session is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Escrow not found or not owned by this member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Evidence is unavailable in the current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /v1/escrows/{id}/disputes:
    post:
      operationId: disputeMemberEscrow
      tags: [Commerce]
      summary: Freeze a member-owned escrow and create a review reference
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "200":
          description: Escrow permanently frozen for review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          description: A valid member session is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Escrow not found or not owned by this member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Dispute is unavailable in the current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
...
  /v1/circles/{circleId}/stories:
    post:
      operationId: createAnansesemStory
      tags: [Community]
      summary: Start a private two-author Anansesɛm relay
      description: The server derives both current authors from an exact two-member circle.
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnansesemCreateInput"
      responses:
        "201":
          description: Private story created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnansesemEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AnansesemNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/stories/{storyId}:
    get:
      operationId: getAnansesemStory
      tags: [Community]
      summary: Read a private story relay
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AnansesemStoryId"
      responses:
        "200":
          description: Privacy-safe story projection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnansesemEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AnansesemNotAvailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/stories/{storyId}/passages:
    post:
      operationId: addAnansesemPassage
      tags: [Community]
      summary: Add the current author's next passage
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AnansesemStoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnansesemPassageMutationInput"
      responses:
        "200":
          description: Passage retained.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnansesemEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AnansesemNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/stories/{storyId}/passages/{passageId}:
    put:
      operationId: editAnansesemPassage
      tags: [Community]
      summary: Edit one passage owned by the current author
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AnansesemStoryId"
        - name: passageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnansesemPassageMutationInput"
      responses:
        "200":
          description: Passage revision retained and publication grants cleared.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnansesemEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AnansesemNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/stories/{storyId}/publication-grants:
    post:
      operationId: grantAnansesemPublication
      tags: [Community]
      summary: Grant publication consent for the current draft fingerprint
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AnansesemStoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnansesemRevisionInput"
      responses:
        "200":
          description: Current-draft grant retained.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnansesemEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AnansesemNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/stories/{storyId}/publish:
    post:
      operationId: publishAnansesemStory
      tags: [Community]
      summary: Publish the current draft after both fingerprint-bound grants
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AnansesemStoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnansesemRevisionInput"
      responses:
        "200":
          description: Redacted edition published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnansesemEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AnansesemNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    MemberBearer:
      type: http
      scheme: bearer
      bearerFormat: Opaque
      description: Short-lived member access token; the session subject must match memberId.
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Staff token carrying the explicitly documented capability and recent-MFA claim where required.
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Safe caller-provided identifier; invalid values are replaced.
      schema:
        $ref: "#/components/schemas/CorrelationId"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key reused for retries of the same command.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
    CircleRoomId:
      name: circleId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    AnansesemStoryId:
      name: storyId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  headers:
    CorrelationId:
      required: true
      description: Effective request correlation identifier.
      schema:
        $ref: "#/components/schemas/CorrelationId"
  responses:
    Unauthorized:
      description: A valid authenticated session is required.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AnansesemNotAvailable:
      description: The story or its exact two-member circle context is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InvalidJSON:
      description: The body is malformed, oversized, contains unknown fields, or contains multiple values.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidJSON:
              value:
                error:
                  code: invalid_json
                  message: The request body must be one valid JSON object.
                meta:
                  correlationId: 01J9EXAMPLE000000000000000
    UnsupportedMediaType:
      description: Content-Type is not application/json.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationFailed:
      description: Request fields or the idempotency key are invalid.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidEmail:
              value:
                error:
                  code: validation_failed
                  message: One or more fields are invalid.
                  details:
                    - field: email
                      reason: must be a valid email address
                meta:
                  correlationId: 01J9EXAMPLE000000000000000
    InternalError:
      description: An unexpected server failure occurred.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    CorrelationId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
    Metadata:
      type: object
      additionalProperties: false
      required: [correlationId]
      properties:
        correlationId:
          $ref: "#/components/schemas/CorrelationId"
...
    AnansesemCreateInput:
      type: object
      additionalProperties: false
      required: [titleCode]
      properties:
        titleCode:
          type: string
          pattern: ^[a-z0-9][a-z0-9_.-]{1,47}$
    AnansesemRevisionInput:
      type: object
      additionalProperties: false
      required: [expectedRevision]
      properties:
        expectedRevision:
          type: integer
          format: int64
          minimum: 1
    AnansesemPassageMutationInput:
      type: object
      additionalProperties: false
      required: [content, expectedRevision]
      properties:
        expectedRevision:
          type: integer
          format: int64
          minimum: 1
        content:
          type: string
          minLength: 1
          maxLength: 1200
    AnansesemPassageData:
      type: object
      additionalProperties: false
      required: [id, ordinal, content, yours, createdAt, editedAt]
      properties:
        id:
          type: string
        ordinal:
          type: integer
          minimum: 0
        content:
          type: string
        yours:
          type: boolean
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    AnansesemEditionPassage:
      type: object
      additionalProperties: false
      required: [ordinal, content]
      properties:
        ordinal:
          type: integer
          minimum: 0
        content:
          type: string
    AnansesemEditionData:
      type: object
      additionalProperties: false
      required: [version, titleCode, passages, publishedAt]
      properties:
        version:
          type: integer
          format: int64
          minimum: 1
        titleCode:
          type: string
        passages:
          type: array
          items:
            $ref: "#/components/schemas/AnansesemEditionPassage"
        publishedAt:
          type: string
          format: date-time
    AnansesemData:
      type: object
      additionalProperties: false
      required:
        [
          id,
          titleCode,
          passages,
          yourTurn,
          yourGrant,
          otherGrant,
          bothGranted,
          editions,
          revision,
        ]
      properties:
        id:
          type: string
        titleCode:
          type: string
        passages:
          type: array
          maxItems: 40
          items:
            $ref: "#/components/schemas/AnansesemPassageData"
        yourTurn:
          type: boolean
        yourGrant:
          type: boolean
        otherGrant:
          type: boolean
        bothGranted:
          type: boolean
        editions:
          type: array
          items:
            $ref: "#/components/schemas/AnansesemEditionData"
        revision:
          type: integer
          format: int64
          minimum: 1
    AnansesemEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AnansesemData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    EscrowMilestoneData:
      type: object
      additionalProperties: false
      required:
        [
          id,
          grossPesewas,
          feePesewas,
          deliveryConfirmed,
          acceptanceConfirmed,
          settled,
        ]
      properties:
        id:
          type: string
        grossPesewas:
          type: integer
          format: int64
          minimum: 1
        feePesewas:
          type: integer
          format: int64
          minimum: 0
        deliveryConfirmed:
          type: boolean
        acceptanceConfirmed:
          type: boolean
        settled:
          type: boolean
        statementRef:
          type: string
...
    EscrowData:
      type: object
      additionalProperties: false
      required:
        [
          escrowId,
          engagementId,
          fundedPesewas,
          settledPesewas,
          termsId,
          termsVersion,
          milestones,
          disputed,
          revision,
        ]
      properties:
        escrowId:
          type: string
        engagementId:
          type: string
        fundedPesewas:
          type: integer
          format: int64
          minimum: 1
        settledPesewas:
          type: integer
          format: int64
          minimum: 0
        termsId:
          type: string
        termsVersion:
          type: integer
          format: int64
          minimum: 1
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/EscrowMilestoneData"
        disputed:
          type: boolean
        escalationRef:
          type: string
        revision:
          type: integer
          format: int64
          minimum: 1
    EscrowEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/EscrowData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    EscrowListData:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EscrowData"
    EscrowListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/EscrowListData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        meta:
          $ref: "#/components/schemas/Metadata"
...